Moving NakadiRecordMapper from core-services to core-common
